+++ A4/Student_Updating.py
@@ -100,9 +100,9 @@ def add_grades_to_student(student_location, students_list):
     """
     while True:
         try:
-            # converts grade input to int
-            grade = int(input('Enter a grade between 0 and 100'
-                              ' to add to student type anything else to stop adding\n').strip())
+            # converts grade input to float
+            grade = float(input('Enter a grade between 0 and 100'
+                                ' to add to student type anything else to stop adding\n').strip())
         except ValueError:
             print('non integer grade entered, exiting edit\n')
             break
